backend/requirements.txt
asgiref==3.9.1
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.6.0
